Correction and typo
Added a link to click the picture to see it big. My site use lighbox
view of images when linked directly to it.

Corrected a bug for the price of one item. The call to the function
missed an argument and made the call unworking

Corrected a typo in archive.php